+++ nemo_curator/stages/text/experimental/translation/evaluation/__init__.py
@@ -1,9 +1,9 @@
-from nemo_curator.stages.text.translation.evaluation.faith import (
+from nemo_curator.stages.text.experimental.translation.evaluation.faith import (
     FAITH_KEYS,
     _SCORE_COLUMNS,
     FaithEvalFilter,
 )
-from nemo_curator.stages.text.translation.evaluation.text_quality import (
+from nemo_curator.stages.text.experimental.translation.evaluation.text_quality import (
     TextQualityMetricStage,
     compute_text_quality_metric,
 )
