+++ nemo_curator/stages/text/experimental/translation/prompts/faith_eval.yaml
