+++ nemo_curator/stages/text/experimental/translation/prompts/translate.yaml
