+++ nemo_curator/stages/text/experimental/translation/utils/__init__.py
@@ -1,17 +1,17 @@
-from nemo_curator.stages.text.translation.utils.async_utils import run_async_safe
-from nemo_curator.stages.text.translation.utils.field_paths import (
+from nemo_curator.stages.text.experimental.translation.utils.async_utils import run_async_safe
+from nemo_curator.stages.text.experimental.translation.utils.field_paths import (
     extract_nested_fields,
     is_wildcard_path,
     normalize_text_field,
     parse_structured_value,
     set_nested_fields,
 )
-from nemo_curator.stages.text.translation.utils.metadata import (
+from nemo_curator.stages.text.experimental.translation.utils.metadata import (
     build_translation_metadata,
     merge_faith_scores_into_metadata,
     reconstruct_messages_with_translation,
 )
-from nemo_curator.stages.text.translation.utils.prompt_loader import load_prompt_template
+from nemo_curator.stages.text.experimental.translation.utils.prompt_loader import load_prompt_template
 
 __all__ = [
     "build_translation_metadata",
